renamed files from featuredApps to featuredCategories
