+++ lib/featuredCategories.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const parseFeaturedApps = require('./utils/parseFeaturedApps');
+const parseFeaturedCategories = require('./utils/parseFeaturedCategories');
 
 function featuredApp (opts) {
   return new Promise(function (resolve, reject) {
@@ -11,7 +11,7 @@ function featuredApp (opts) {
       num: 100
     }, opts);
 
-    parseFeaturedApps(opts)
+    parseFeaturedCategories(opts)
     .then(resolve)
     .catch(reject);
   });
