+++ lib/mappers/featuredCategories.js
