+++ lib/requesters/featuredCategoriesMappedRequests.js
@@ -5,7 +5,7 @@ const R = require('ramda');
 const request = require('../utils/request');
 const scriptData = require('../utils/scriptData');
 const featuredAppList = require('../utils/featuredAppList');
-const { REQUEST_MAPPINGS } = require('../mappers/featuredApps');
+const { REQUEST_MAPPINGS } = require('../mappers/featuredCategories');
 const { BASE_URL } = require('../utils/configurations');
 
 function getBodyForRequests ({
