+++ lib/utils/parseFeaturedCategories.js
@@ -1,14 +1,14 @@
 'use strict';
 
-const debug = require('debug')('google-play-scraper:parseFeaturedApps');
+const debug = require('debug')('google-play-scraper:parseFeaturedCategories');
 const request = require('./request');
 const R = require('ramda');
 const scriptData = require('./scriptData');
-const { processAndRecur } = require('../requesters/featuredAppsMappedRequests');
+const { processAndRecur } = require('../requesters/featuredCategoriesMappedRequests');
 const { CLUSTER_BASE_URL } = require('./configurations');
-const { INITIAL_MAPPINGS } = require('../mappers/featuredApps');
+const { INITIAL_MAPPINGS } = require('../mappers/featuredCategories');
 
-function parseFeaturedApps (opts) {
+function parseFeaturedCategories (opts) {
   
   const url = `${CLUSTER_BASE_URL}?gl=${opts.country}&hl=${opts.lang}`;
   
@@ -24,4 +24,4 @@ function parseFeaturedApps (opts) {
     .catch(console.error);
 }
 
-module.exports = parseFeaturedApps;
+module.exports = parseFeaturedCategories;
